+++ Sources/Animators/TransitionAnimator/PortalAnimator.swift
@@ -41,6 +41,11 @@ extension PortalAnimator: UIViewControllerAnimatedTransitioning {
       return
     }
 
+    let (_, tempToViewController, _) = retrieveViewControllers(transitionContext: transitionContext)
+    if let toViewController = tempToViewController {
+      toView.frame = transitionContext.finalFrame(for: toViewController)
+    }
+
     switch fromDirection {
     case .forward:
       executeForwardAnimation(transitionContext: transitionContext, containerView: containerView, fromView: fromView, toView: toView)