+++ CHANGELOG.md
@@ -26,6 +26,7 @@ N/A
 
 - Fix some activity indicator animations . [#562](https://github.com/IBAnimatable/IBAnimatable/pull/562) by [@phimage](https://github.com/phimage)
 - Fix wrong setted center in UIView after using NatGeo transition. [#578](https://github.com/IBAnimatable/IBAnimatable/pull/578) by [@edon2005](https://github.com/edon2005) 
+- Fix destroyed constraints in view of toViewController after transition on iOS 9/10. [#580](https://github.com/IBAnimatable/IBAnimatable/pull/580) by [@edon2005](https://github.com/edon2005)
 
 ### [5.1.0](https://github.com/IBAnimatable/IBAnimatable/releases/tag/5.1.0)